requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
cors==1.0.1
python-multipart==0.0.6
